grpc: Remove JVM Generator related code
