+++ gradle-plugin/src/main/kotlin/kotlinx/rpc/proto/DefaultProtoSourceSet.kt
@@ -96,13 +96,11 @@ internal fun Project.createProtoExtensions() {
     }
 
     project.withKotlinKmpExtension {
-        findOrCreateAndConfigure("jvmMain", null)
-        findOrCreateAndConfigure("jvmTest", null)
         findOrCreateAndConfigure("commonMain", null)
         findOrCreateAndConfigure("commonTest", null)
 
         sourceSets.configureEach {
-            if (name == "jvmMain" || name == "jvmTest" || name == "commonMain" || name == "commonTest") {
+            if (name == "commonMain" || name == "commonTest") {
                 findOrCreateAndConfigure(name, this)
             }
         }
