+++ gradle-plugin/src/test/resources/projects/GrpcKmpProjectTest/main_and_test_mixed_sources/src/commonTest/proto/other.proto
