+++ gradle-plugin/src/test/resources/projects/GrpcKmpProjectTest/minimal_grpc_configuration/src/commonMain/proto/some.proto
