+++ gradle-plugin/src/test/resources/projects/GrpcKmpProjectTest/no_grpc/build.gradle.kts
@@ -11,6 +11,6 @@ plugins {
 
 // should nonetheless be available
 protoSourceSets {
-    jvmMain {}
-    jvmTest {}
+    commonMain {}
+    commonTest {}
 }
