+++ gradle-plugin/src/test/resources/projects/GrpcKmpProjectTest/no_jvm_targets/build.gradle.kts
@@ -16,6 +16,6 @@ kotlin {
 
 // should nonetheless be available
 protoSourceSets {
-    jvmMain {}
-    jvmTest {}
+    commonMain {}
+    commonTest {}
 }
