+++ gradle-plugin/src/test/resources/projects/GrpcKmpProjectTest/test_only_sources/src/commonTest/proto/some.proto
