+++ gradle-plugin/src/test/resources/projects/GrpcKmpProjectTest/with_other_targets/build.gradle.kts
@@ -16,8 +16,8 @@ kotlin {
 }
 
 protoSourceSets {
-    jvmMain {}
-    jvmTest {}
+    commonMain {}
+    commonTest {}
 }
 
 rpc {
