+++ gradle-plugin/src/test/resources/projects/GrpcKmpProjectTest/with_other_targets/src/commonMain/proto/some.proto
