+++ grpc/grpc-core/src/commonTest/proto/exclude/empty_deprecated.proto
