+++ grpc/grpc-core/src/commonTest/proto/exclude/enum.proto
