+++ grpc/grpc-core/src/commonTest/proto/exclude/enum_options.proto
