+++ grpc/grpc-core/src/commonTest/proto/exclude/example.proto
