+++ grpc/grpc-core/src/commonTest/proto/exclude/image-recognizer.proto
