+++ grpc/grpc-core/src/commonTest/proto/exclude/multiple_files.proto
