+++ grpc/grpc-core/src/commonTest/proto/exclude/one_of.proto
