+++ grpc/grpc-core/src/commonTest/proto/exclude/primitive_service.proto
