+++ grpc/grpc-core/src/commonTest/proto/exclude/reference_service.proto
