+++ grpc/grpc-core/src/commonTest/proto/exclude/streaming.proto
