+++ grpc/grpc-core/src/commonTest/proto/exclude/test_map.proto
